Restructured operators and fix data processing logic
