+++ minference/dist_ops/minfer_zigzag.py
@@ -8,9 +8,9 @@ from .utils import (
     RingComm,
     shuffle_zigzag_input, recover_zigzag_output,
 )
-from minference.ops.utils import build_index, convert_blockmask
-from minference.ops.minference_attn_triton import block_bar_attn_fwd
-from minference.ops.minference_attn import block_attn_bwd, bar_attn_bwd
+from minference.ops.op_utils.vertical_slash_utils import build_index, convert_blockmask
+from minference.ops.pit_sparse_flash_attention_v3_triton import block_bar_attn_fwd
+from minference.ops.pit_sparse_flash_attention_v3 import block_attn_bwd, bar_attn_bwd
 
 def minfer_zigzag_forward(
     process_group: dist.ProcessGroup,
