+++ minference/ops/pit_sparse_flash_attention_v3.py
@@ -24,7 +24,7 @@ if torch.version.hip is None:
         sys.setdlopenflags(original_flags)
     # NOTE: Block-Sparse-Attention/csrc/block_sparse_attn/src/flash_blockmask.h: add head_idx to blockmask_ptr
 
-from .utils import build_index_local
+from .op_utils.vertical_slash_utils import build_index_local
 
 
 def block_attn_fwd(
@@ -234,164 +234,6 @@ def _triton_bar_attn_fwd_kernel(
     tl.store(o_ptrs, acc.to(Out.type.element_ty), mask=m_mask[:, None])
     tl.store(lse_ptrs, s, mask=(m_mask & overflow_mask))
 
-def stable_sigmoid(x: torch.Tensor):
-    return torch.where(
-        x >= 0,
-        1 / (1 + torch.exp(-x)),
-        torch.exp(x) / (1 + torch.exp(x))
-    )
-
-def naive_bar_attn_fwd(
-        q, k, v, 
-        sm_scale, 
-        bar_cnt, bar_idx, 
-        out, softmax_lse,
-        step, granularity, BLOCK_N=64
-    ):
-    """
-    Naive PyTorch implementation of the Triton bar attention forward kernel.
-    
-    Args:
-      q: Query tensor of shape [B, num_qo_heads, num_tokens, head_dim]
-      k: Key tensor of shape [B, num_kv_heads, num_tokens,  head_dim]
-      v: Value tensor of shape [B, num_kv_heads, num_tokens, head_dim]
-      sm_scale: A scalar (float) softmax scale.
-      bar_cnt: Tensor of shape [B, num_qo_heads, num_blocks, world_size+1]
-               where each block (row) holds bar boundary indices.
-      bar_idx: Tensor of shape [B, num_qo_heads, num_blocks, nnz_v]
-               containing indices of keys (columns) for each block.
-      out: Output tensor of shape [B, num_qo_heads, num_tokens, head_dim].
-           This is assumed to have a previous value to merge with.
-      softmax_lse: Tensor of shape [B, num_qo_heads, num_tokens] containing
-                   the previous log-sum-exp values.
-      step: integer step indicating which pair of boundaries to use in bar_cnt.
-      granularity: BLOCK_M, i.e. the number of query tokens processed per block.
-      BLOCK_N: Block size for the key dimension (default: 64)
-      
-    This function updates `out` and `softmax_lse` in-place.
-    """
-    # Get dimensions from q.
-    B, num_tokens, num_qo_heads, head_dim = q.shape
-
-    # Determine number of query blocks (each corresponding to a row in bar_cnt/bar_idx).
-    num_blocks = math.ceil(num_tokens / granularity)
-
-    # Compute the ratio for mapping query-head to key/value head.
-    head_ratio = num_qo_heads // k.shape[2]  # since k.shape[1] is num_kv_heads
-
-    # Precompute scale for q: note that 1.44269504 = log2(e)
-    qk_scale = sm_scale * 1.44269504
-
-    ln2 = 0.69314718  # constant for converting exp2 to exp
-
-    # Loop over batch and query-head
-    for b in range(B):
-        for qh in range(num_qo_heads):
-            # corresponding key/value head index
-            kvh = qh // head_ratio
-
-            # Loop over query blocks (rows)
-            for block in range(num_blocks):
-                start_m = block * granularity
-                end_m = min(start_m + granularity, num_tokens)
-                block_size = end_m - start_m
-
-                # Get bar boundaries for this block & step:
-                # bar_cnt is assumed to store cumulative indices per block.
-                bar_l = bar_cnt[b, qh, block, step].item()   # starting index (inclusive)
-                bar_r = bar_cnt[b, qh, block, step + 1].item() # ending index (exclusive)
-                if bar_l >= bar_r:
-                    continue  # nothing to do in this block
-
-                # Initialize accumulators per query token in the block.
-                # m_i tracks the running maximum (in "log2" domain).
-                m_i = torch.full((block_size,), -float('inf'), device=q.device, dtype=torch.float32)
-                # l_i tracks the running sum-of-weights.
-                l_i = torch.zeros(block_size, device=q.device, dtype=torch.float32)
-                # acc accumulates the weighted sum of values.
-                acc = torch.zeros((block_size, head_dim), device=q.device, dtype=torch.float32)
-
-                # Load and scale the q block.
-                # Shape: [block_size, head_dim]
-                q_block = q[b, start_m:end_m, qh, :] * qk_scale
-
-                # Loop over key indices in steps of BLOCK_N
-                for n_start in range(bar_l, bar_r, BLOCK_N):
-                    n_end = min(n_start + BLOCK_N, bar_r)
-                    
-                    # Load column indices from bar_idx.
-                    # bar_idx shape: [nnz_v] for this block.
-                    cols = bar_idx[b, qh, block, n_start:n_end]
-                    cols = cols.long()
-                    
-                    k_selected = k[b, cols, kvh, :]  # shape: [n_valid, head_dim]
-                    v_selected = v[b, cols, kvh, :]  # shape: [n_valid, head_dim]
-
-                    # Compute scaled dot product: [block_size, head_dim] x [head_dim, n_valid]
-                    # Result: [block_size, n_valid]
-                    qk = torch.matmul(q_block, k_selected.T)
-
-                    # Numerically stable softmax update in the log2 domain.
-                    # m_i_new = max(m_i, max(qk, dim=1))
-                    cur_max, _ = qk.max(dim=1)
-                    m_i_new = torch.max(m_i, cur_max)
-
-                    alpha = torch.exp2((m_i - m_i_new))
-                    p = torch.exp2((qk - m_i_new.unsqueeze(1)))
-                    
-                    # Update acc and l_i.
-                    # Scale previous acc by alpha.
-                    acc = acc * alpha.unsqueeze(1) + torch.matmul(p.to(q.dtype), v_selected)
-
-                    l_i = l_i * alpha + p.sum(dim=1)
-
-                    # Update m_i to the new maximum.
-                    m_i = m_i_new
-
-                # check zeros in l_i, if any, print out the indices
-                if (l_i == 0).any():
-                    zero_indices = torch.nonzero(l_i == 0).squeeze()
-                    print(f"Rank {dist.get_rank()} | Zeros in l_i (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {zero_indices}")
-
-                # Finalize the block output.
-                # Compute weighted output.
-                acc_1 = acc / l_i.unsqueeze(1)
-                s_1 = m_i * ln2 + torch.log(l_i)
-                # check positive infinity in s_1, if any, print out the indices
-                if torch.isinf(s_1).any() and ( torch.isinf(s_1) & (s_1 > 0) ).any():
-                    mask = torch.isinf(s_1) & (s_1 > 0)
-                    print(f"Rank {dist.get_rank()} | Positive infinity in s_1 (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {torch.nonzero(mask).squeeze()}")
-                # check negative infinity in s_1, if any, print out the indices
-                if torch.isinf(s_1).any() and ( torch.isinf(s_1) & (s_1 < 0) ).any():
-                    mask = torch.isinf(s_1) & (s_1 < 0)
-                    print(f"Rank {dist.get_rank()} | Negative infinity in s_1 (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {torch.nonzero(mask).squeeze()}")
-
-                # Load previous stored values (accumulated output and LSE).
-                old_out = out[b, start_m:end_m, qh, :].to(acc_1.dtype)
-                old_lse = softmax_lse[b, qh, start_m:end_m]
-                # check positive infinity in old_lse, if any, print out the indices
-                if torch.isinf(old_lse).any() and ( torch.isinf(old_lse) & (old_lse > 0) ).any():
-                    mask = torch.isinf(old_lse) & (old_lse > 0)
-                    print(f"Rank {dist.get_rank()} | Positive infinity in old_lse (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {torch.nonzero(mask).squeeze()}")
-
-                # -------------------------------------------------
-                # Logsigmoid solution
-                # out - old_out, block_out - acc1, lse - old_lse, block_lse - s_1
-                new_out = old_out - F.sigmoid(s_1 - old_lse).unsqueeze(1) * (old_out - acc_1)
-                new_lse = s_1 - F.logsigmoid(s_1 - old_lse)
-                if torch.isinf(new_lse).any() and ( torch.isinf(new_lse) & (new_lse > 0) ).any():
-                    mask = torch.isinf(new_lse) & (new_lse > 0)
-                    print(f"Rank {dist.get_rank()} | Positive infinity in new_lse (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {torch.nonzero(mask).squeeze()}")
-
-                    pos_inf_indices = torch.nonzero(mask).squeeze()
-                    print(f"Rank {dist.get_rank()} | Values of (old_lse - s_1) resulting in pos-inf in theta (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {(old_lse - s_1)[pos_inf_indices]}")
-                    print(f"Rank {dist.get_rank()} | Values of (old_lse) resulting in pos-inf in theta (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {(old_lse)[pos_inf_indices]}")
-                    print(f"Rank {dist.get_rank()} | Values of (s_1) resulting in pos-inf in theta (step = {step}, batch_idx={b}, head_idx={qh}, block={block}): {(s_1)[pos_inf_indices]}")
-
-                # Store back into out and softmax_lse.
-                out[b, start_m:end_m, qh, :] = new_out.to(out.dtype)
-                softmax_lse[b, qh, start_m:end_m] = new_lse
-    return out, softmax_lse
 
 
 def bar_attn_fwd(
