+++ mtraining/requirements.txt
@@ -1,18 +1,11 @@
 transformers==4.48.0
 datasets==2.20.0
 tensorboard
-scikit-learn 
-matplotlib
-seaborn
 jieba 
 rouge 
 nltk 
 rouge_score 
 evaluate
-triton==3.0.0
 
-mosaicml-cli==0.5.34
+# For Data Preparation 
 mosaicml-streaming==0.8.1
-sentencepiece==0.1.99
-tiktoken==0.7.0
-zstandard==0.22.0
